requirements.txt
# Dependencies for NeuroSniper trading system
# Supports 49 features (including Safe Mode) for intraday trading with Python 3.12.7
# Format: package==version  # Purpose

requests==2.32.3        # HTTP requests for APIs
pandas==2.2.3           # Data manipulation
numpy==1.26.4           # Numerical computations (compatible with tensorflow)
matplotlib==3.9.2       # Dashboard graphs
seaborn==0.13.2         # Enhanced visualizations
scikit-learn==1.5.2     # ML models
tensorflow==2.17.0      # Deep learning (LSTM)
joblib==1.4.2           # Model serialization
nsepython==2.8          # NSE API for OI/PCR data
ta==0.11.0              # Technical analysis indicators
smartapi-python==1.5.5  # Angel One API for market data and trading
websocket-client==1.8.0 # WebSocket for market data
pyotp==2.9.0            # OTP for Angel One API
pyyaml==6.0.2           # YAML config parsing
pydantic==2.9.2         # Config validation
jsonschema==4.23.0      # Alternative validation
logzero==1.7.0          # Logging utility
watchdog==5.0.3         # File/system monitoring (Safe Mode crash protection)
psutil==6.0.0           # System resource monitoring
python-telegram-bot==21.6 # Telegram bot API (Safe Mode alerts)
streamlit==1.39.0       # Streamlit dashboard (Safe Mode indicator)
pytest==8.3.3           # Unit testing
pytest-cov==5.0.0       # Test coverage
pytest-asyncio==0.24.0  # Async testing support
pylint==3.3.1           # Linting
black==24.8.0           # Code formatting
isort==5.13.2           # Import sorting
speedtest-cli==2.1.3    # Internet health check for Safe Mode
# sqlite3               # SQLite (standard library, no installation needed)